fix: prevent AttributeError in head() and add regression test
